+++ rusk/src/lib/http/rusk.rs
@@ -30,6 +30,7 @@ impl HandleRequest for Rusk {
         #[allow(clippy::match_like_matches_macro)]
         match request.uri.inner() {
             ("contracts", Some(_), _) => true,
+            ("driver", Some(_), _) => true,
             ("node", _, "provisioners") => true,
             ("account", Some(_), "status") => true,
             ("node", _, "crs") => true,
@@ -50,6 +51,9 @@ impl HandleRequest for Rusk {
                     feeder,
                     request.is_json(),
                 )
+            }
+            ("driver", Some(contract_id), method) => {
+                Self::handle_data_driver(contract_id, method, &request.data)
             }
             ("node", _, "provisioners") => self.get_provisioners(),
 
@@ -78,6 +82,44 @@ impl Rusk {
             _ => None,
         };
         Ok(driver)
+    }
+
+    fn handle_data_driver(
+        contract_id: &str,
+        method: &str,
+        data: &RequestData,
+    ) -> anyhow::Result<ResponseData> {
+        let (method, target) = method.split_once(':').unwrap_or((method, ""));
+        let driver = Self::data_driver(contract_id.to_string())?
+            .ok_or(anyhow::anyhow!("Unsupported contractId {contract_id}"))?;
+        let result = match method {
+            "decode_event" => ResponseData::new(
+                driver
+                    .decode_event(target, data.as_bytes())
+                    .map_err(|e| anyhow::anyhow!("{e}"))?,
+            ),
+            "decode_input_fn" => ResponseData::new(
+                driver
+                    .decode_input_fn(target, data.as_bytes())
+                    .map_err(|e| anyhow::anyhow!("{e}"))?,
+            ),
+            "decode_output_fn" => ResponseData::new(
+                driver
+                    .decode_output_fn(target, data.as_bytes())
+                    .map_err(|e| anyhow::anyhow!("{e}"))?,
+            ),
+            "encode_input_fn" => ResponseData::new(
+                driver
+                    .encode_input_fn(target, &data.as_string())
+                    .map_err(|e| anyhow::anyhow!("{e}"))?,
+            ),
+            "get_schema" => ResponseData::new(driver.get_schema().to_string()),
+            "get_version" => {
+                ResponseData::new(driver.get_version().to_string())
+            }
+            method => anyhow::bail!("Unsupported datadriver method {method}"),
+        };
+        Ok(result)
     }
 
     fn handle_contract_query(