+++ rusk/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `Content-Type: application/json` support for `/on/contracts` endpoint
 - Add transaction serialization check
 - Add max transaction size check
+- Add `/on/driver:<contract>/<method>:<target>` endpoint
 
 ## [1.2.0] - 2025-03-20